Fixed doors and adjusted hallway & enemy parameters
Fixed so the first door of a room is always the room you came from, the
next door is the next room, and any other doors are random links to
other rooms. Had to implement and entrance room for that.

Changed the likely hood of getting a hallway and enemy and established
global variables to manage further adjustment.